Move tests of ApplicationFiles to a dedicated file
